Fix overflows

+++ rust/src/p0007.rs
@@ -12,7 +12,7 @@ What is the 10 001st prime number?
 use crate::primes::primes;
 
 pub fn p0007() -> i128 {
-    return primes::<u32>()
+    return primes::<u64>()
         .take(10001)
         .last()
         .unwrap()
